% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{ADH}
\alias{ADH}
\title{Dataset from Autor, Dorn and Hanson (2013)}
\format{An object of class \code{list} of length 3.}
\source{
We thank David Dorn for helping us with the construction of the share
    matrix. The remaining data was obtained from David Dorn's website,
    \url{http://ddorn.net/data.htm}.
}
\usage{
ADH
}
\description{
Subset of data from Autor, Dorn and Hanson (2013, ADH) that is used to
illustrate the confidence intervals implemented in this package.
}
\references{
{

\cite{Autor, David H., David Dorn, and Gordon H. Hanson, "The China syndrome:
Local labor market effects of import competition in the United States,"
American Economic Review, 2013, 103 (6), 2121–68.}

\cite{Adão, Rodrigo, Kolesár, Michal, and Morales, Eduardo,
"Shift-Share Designs: Theory and Inference", 2018, arXiv:1806.07928}

}
}
\keyword{datasets}
