% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{ADH_W}
\alias{ADH_W}
\title{Weight matrix variable from Autor, Dorn and Hanson (2013)}
\format{A matrix with dimensions 1,444 x 775}
\source{
David Dorn's website \url{http://ddorn.net/data.htm}
}
\usage{
ADH_W
}
\description{
Weight matrix variable from Autor, Dorn and Hanson (2013)
}
\references{
{

\cite{Autor, David H., David Dorn, and Gordon H. Hanson, "The China syndrome:
Local labor market effects of import competition in the United States,"
American Economic Review, 2013, 103 (6), 2121–68.}

}
}
\keyword{datasets}
