% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{ADH_Xs}
\alias{ADH_Xs}
\title{Sector shock variable from Autor, Dorn and Hanson (2013)}
\format{A vector of length 775}
\source{
TODO
}
\usage{
ADH_Xs
}
\description{
Sector shock variable from Autor, Dorn and Hanson (2013)
}
\references{
{

\cite{Autor, David H., David Dorn, and Gordon H. Hanson, "The China syndrome:
Local labor market effects of import competition in the United States,"
American Economic Review, 2013, 103 (6), 2121–68.}

}
}
\keyword{datasets}
