% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.R
\docType{data}
\name{BartikSE}
\alias{BartikSE}
\title{Inference in Bartik designs}
\format{An object of class \code{NULL} of length 0.}
\usage{
BartikSE
}
\description{
Compute standard erorrs, confidence intervals, and p-values in Bartik designs
using different inference methods, as specified by \code{method}
}
\keyword{datasets}
