% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.R
\name{ivBartik}
\alias{ivBartik}
\title{Inference in an IV regression with Bartik structure}
\usage{
ivBartik(formula, data, subset, weights, Xs, W, method, beta0 = 0,
  alpha = 0.05, region_cvar = NULL)
}
\arguments{
\item{formula}{object of class \code{"formula"} (or one that can be coerced
to that class) of the form \code{outcome ~ controls | endogenous_regressor}}

\item{data}{optional data frame, list or environment (or object coercible by
\code{as.data.frame} to a data frame) containing the outcome and running
variables in the model. If not found in \code{data}, the variables are
taken from \code{environment(formula)}, typically the environment from
which the function is called.}

\item{subset}{optional vector specifying a subset of observations to be used
in the fitting process.}

\item{weights}{an optional vector of weights to be used in the fitting
process. Should be \code{NULL} or a numeric vector. If non-\code{NULL},
weighted least squares is used with weights \code{weights} (that is,
minimizing \code{sum(weights*residuals^2)}); otherwise ordinary least squares is used.}

\item{Xs}{A vector of sector-level shocks.}

\item{W}{A matrix of sector shares, so that \code{W[i, s]} corresponds to
share of sector \code{s} in region \code{i}}

\item{method}{Vector specifying which inference methods to use:

\describe{

\item{\code{"homosk"}}{Assume i.i.d. homoskedastic errors}

\item{\code{"ehw"}}{Eicker-Huber-White standard errors}

\item{\code{"region_cluster"}}{Clustered standard errors at regional level}

\item{\code{"akm"}}{Adao-Kolesar-Morales}

\item{\code{"akmnull"}}{Adao-Kolesar-Morales with null imposed}

\item{\code{"all"}}{All of the methods above}},}

\item{beta0}{null that is tested (for p-values)}

\item{alpha}{Significance level (confidence intervals will have coverage
\code{1-alpha})}

\item{region_cvar}{A vector of cluster variables, for method \code{cluster_region}.
If the vector \code{1:N} is used, clustering is effectively equivalent to
\code{ehw}}
}
\description{
Inference in an IV regression with Bartik structure
}
\section{Note}{

\code{subset} is evaluated in the same way as variables in \code{formula},
that is first in \code{data} and then in the environment of \code{formula}.
}

