% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{ADH}
\alias{ADH}
\title{Dataset from Autor, Dorn and Hanson (2013)}
\format{A data frame with 1,444 rows and 22 variables:

\describe{
\item{d_sh_empl}{}
\item{d_sh_empl_mfg}{}
\item{d_sh_empl_nmfg}{}
}}
\source{
TODO
}
\usage{
ADH
}
\description{
Subset of ADH data.
}
\references{
{

\cite{Autor, David H., David Dorn, and Gordon H. Hanson, "The China syndrome:
Local labor market effects of import competition in the United States,"
American Economic Review, 2013, 103 (6), 2121–68.}

}
}
\keyword{datasets}
