% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.R
\name{lmBartik.fit}
\alias{lmBartik.fit}
\title{Inference in a regression with Bartik structure}
\usage{
lmBartik.fit(y, Xs, W, Z, beta0 = 0, CI = FALSE, alpha = 0.05,
  clustervar = NULL, method = c("akm", "akm0"))
}
\arguments{
\item{y}{Outcome variable, vector of length \code{N}}

\item{Xs}{Sector-level shocks, vector of length \code{S}}

\item{W}{matrix of sectoral weights with dimensions \code{N}-by-\code{S}}

\item{Z}{Matrix of regional controls, matrix with \code{N} rows}

\item{CI}{For method \code{akmnull} compute confidence
intervals by inverting the tests?}

\item{clustervar}{A vector of cluster variables, for method \code{cluster}.
If the vector \code{1:N} is used, clustering is effectively equivalent to
\code{ehw}}

\item{method}{Vector specifying which inference methods to use:

\describe{

\item{\code{"homosk"}}{Assume i.i.d. homoskedastic errors}

\item{\code{"ehw"}}{Eicker-Huber-White standard errors}

\item{\code{"cluster"}}{Clustered standard errors}

\item{\code{"akm"}}{Adao-Kolesar-Morales}

\item{\code{"akmnull"}}{Adao-Kolesar-Morales with null imposed}},}
}
\description{
Basic computing engine to calculate, confidence intervals, and p-values in
Bartik designs using different inference methods, as specified by
\code{method}
}
