% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.R
\name{lmBartik.fit}
\alias{lmBartik.fit}
\title{Inference in a regression with Bartik structure}
\usage{
lmBartik.fit(y, Xs, W, Z, w = NULL, method = c("akm", "akm0"), beta0 = 0,
  alpha = 0.05, region_cvar = NULL, residual_sector = FALSE)
}
\arguments{
\item{y}{Outcome variable, vector of length \code{N}}

\item{Xs}{A vector of sector-level shocks.}

\item{W}{A matrix of sector shares, so that \code{W[i, s]} corresponds to
share of sector \code{s} in region \code{i}}

\item{Z}{Matrix of regional controls, matrix with \code{N} rows}

\item{w}{vector of weights (length \code{N}) to be used in the fitting
process. If not \code{NULL}, weighted least squares is used with weights
\code{w}, i.e., \code{sum(w * residuals^2)} is minimized.}

\item{method}{Vector specifying which inference methods to use:

\describe{

\item{\code{"homosk"}}{Assume i.i.d. homoskedastic errors}

\item{\code{"ehw"}}{Eicker-Huber-White standard errors}

\item{\code{"region_cluster"}}{Clustered standard errors at regional level}

\item{\code{"akm"}}{Adao-Kolesar-Morales}

\item{\code{"akmnull"}}{Adao-Kolesar-Morales with null imposed}

\item{\code{"all"}}{All of the methods above}},}

\item{beta0}{null that is tested (for p-values)}

\item{alpha}{Significance level (confidence intervals will have coverage
\code{1-alpha})}

\item{region_cvar}{A vector of cluster variables, for method
\code{cluster_region}. If the vector \code{1:N} is used, clustering is
effectively equivalent to \code{ehw}}

\item{residual_sector}{create a dummy residual sector so weights sum to one.}
}
\description{
Basic computing engine to calculate, confidence intervals, and p-values in
Bartik designs using different inference methods, as specified by
\code{method}
}
